Capstone-2
This company is a product based organization which uses Docker for their containerization needs
within the company. The final product received a lot of traction in the first few
weeks of launch. Now with the increasing demand, the organization needs to
have a platform for automating deployment, scaling and operations of application
containers across clusters of hosts. As a DevOps Engineer, you need to
implement a DevOps lifecycle such that all the requirements are implemented
without any change in the Docker containers in the testing environment.